requirements.txt
Flask==2.3.2
tensorflow==2.14.0
opencv-python==4.9.0.80
numpy==1.23.5
matplotlib==3.7.2
Pillow==10.2.0
rembg==2.0.50
ultralytics==8.0.200
scikit-image==0.21.0
scikit-learn==1.3.2
joblib==1.3.2
gunicorn==21.2.0
